报备后添加隔离记录

+++ community-business/community-business-epidemic/src/main/java/com/roydon/business/epidemic/domain/entity/EpidemicIsolationPolicy.java
@@ -27,7 +27,7 @@ public class EpidemicIsolationPolicy extends BaseEntity {
     /**
      * 隔离天数
      */
-    private Long isolationDay;
+    private Integer isolationDay;
 
 
 }
